+++ src/lib.rs
@@ -60,6 +60,7 @@ mod ike;
 mod inkling;
 mod jack;
 mod kamui;
+mod ken;
 mod kirby;
 mod koopa;
 mod koopajr;
@@ -191,6 +192,7 @@ pub extern "C" fn main() {
 	jack::install();
 	
 	kamui::install();
+	ken::install();
 	kirby::install();
 	koopa::install();
 	koopajr::install();
